fix record shift in xx_*.effective file
File xx_*.effective contains only a sub-set of full control-adjustment time-series,
that corresponds to this simulation. This modification fixes the shift in record-number,
which was ignored in master branch and was not working for all cases in previous version.

Also enable some debug print (if debugLevel >= 3) that were un-usable before (ECCO_VERBOSE
never defined since ECCO_OPTIONS.h was not included).